Imported libsasl2-2.xml
<?xml version="1.0" ?>
<?xml-stylesheet type='text/xsl' href='interface.xsl'?>
<interface uri="http://0install.net/2007/3rd-party/libsasl2-2.xml" xmlns="http://zero-install.sourceforge.net/2004/injector/interface">
  <name>libsasl2-2</name>
  <summary>Cyrus SASL - authentication abstraction library</summary>
  <description>This is the Cyrus SASL API implementation, version 2.1.

SASL is the Simple Authentication and Security Layer, a method for
adding authentication support to connection-based protocols. To use
SASL, a protocol includes a command for identifying and
authenticating a user to a server and for optionally negotiating
protection of subsequent protocol interactions. If its use is
negotiated, a security layer is inserted between the protocol and the
connection. See RFC 2222 for more information.

Any of: ANONYMOUS, CRAM-MD5, DIGEST-MD5, GSSAPI (MIT or Heimdal
Kerberos 5), NTLM, OTP, PLAIN, or LOGIN can be used. If you intend to
use this package on a server that provides SASL authentication, then
you must install some of the libsasl2-modules* packages.</description>
  <package-implementation package="libsasl2-2"/>
  <group arch="Linux-i386">
    <requires interface="http://0install.net/2007/3rd-party/libdb4.6.xml">
      <environment insert="usr/lib" name="LD_LIBRARY_PATH"/>
    </requires>
    <implementation id="sha1new=eb8215753fa6ed28bab588f8e6b59f75ea621ddd" released="2009-09-25" version="2.1.22">
      <archive href="http://ftp.debian.org/debian/pool/main/c/cyrus-sasl2/libsasl2-2_2.1.22.dfsg1-23+lenny1_i386.deb" size="105514"/>
    </implementation>
    <implementation arch="Linux-x86_64" id="sha1new=ce00944b323777c892c8e6dc8e7bee28072f3104" released="2009-09-25" version="2.1.22">
      <archive href="http://ftp.debian.org/debian/pool/main/c/cyrus-sasl2/libsasl2-2_2.1.22.dfsg1-23+lenny1_amd64.deb" size="114804"/>
    </implementation>
  </group>

</interface>
<!-- Base64 Signature
iEYEABECAAYFAkq9DooACgkQBSn9tx+xORBf7wCfSLx+e7p3kAoFuNIoyig4RQXa/LkAn1G2WlaM
kyJ6CZCK4tsyMo+AqHp+

-->