Imported XdgUtils.xml
<?xml version="1.0" ?>
<?xml-stylesheet type='text/xsl' href='interface.xsl'?>
<interface uri="http://0install.net/2007/3rd-party/XdgUtils.xml" xmlns="http://zero-install.sourceforge.net/2004/injector/interface">
  <name>XdgUtils</name>
  <summary>cross-desktop integration tools</summary>
  <description>
Xdg-utils is a set of command line tools that assist applications with a variety
of desktop integration tasks. About half of the tools focus on tasks commonly
required during the installation of a desktop application and the other half
focuses on integration with the desktop environment while the application is
running. Even if the desktop components of your application are limited to an
installer, configuration or management tool, Xdg-utils provides you with an easy
way to enhance the usage experience of your customers by improving the
integration of these components in the user's environment. Best of all, Xdg-
utils is provided as open source and free of charge.</description>
  <homepage>http://portland.freedesktop.org</homepage>
  <group license="OSI Approved :: MIT/X Consortium License">
    <implementation id="sha1new=87a7e092d0a4ec9f02235397c87b454afb806f81" released="2007-03-22" stability="stable" version="1.0.1">
    <archive extract="xdg-utils-1.0.1" href="http://portland.freedesktop.org/download/xdg-utils-1.0.1-scripts.tgz" size="47983" type="application/x-compressed-tar"/>
  </implementation>
    <implementation id="sha1new=4372620f1b806ff9c7ee35f57cd43283e4912406" released="2007-06-25" stability="stable" version="1.0.2">
    <archive extract="xdg-utils-1.0.2" href="http://portland.freedesktop.org/download/xdg-utils-1.0.2-scripts.tgz" size="47751" type="application/x-compressed-tar"/>
  </implementation>
  </group>
</interface>
<!-- Base64 Signature
iD8DBQBGqyGbBSn9tx+xORARAt8LAKCHBQSEF49gd7Rla+GJ6jszKOjr0wCeICTjOm+D6zZW+8Ux
JK0BygHJt2U=

-->